refactor data preprocessing, fix mllm rlhf
